Added shared crate for contracts
