+++ contracts/Cargo.toml
@@ -26,6 +26,9 @@ basic-analyzer = { path = "contracts/basic-analyzer", version = "0.1.0", feature
 ethereum-analyzer = { path = "contracts/ethereum-analyzer", version = "0.1.0", features = [
   "library",
 ] }
+contracts-core = { path = "contracts/contracts-core", version = "0.1.0", features = [
+  "library",
+] }
 
 [profile.release.package.sample]
 codegen-units = 1
