+++ contracts/contracts/contracts-core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "osmosis-analyzer"
+name = "contracts-core"
 version.workspace = true
 edition = "2021"
 description = ""
@@ -16,18 +16,11 @@ library = []
 test-utils = []
 
 [dependencies]
-contracts-core = { workspace = true }
 cosmwasm-schema = { workspace = true }
-cw2 = { workspace = true }
-bindings = { workspace = true }
 cosmwasm-std = { workspace = true, features = ["staking"] }
-schemars = { workspace = true }
+cw2 = { workspace = true }
 serde = { workspace = true }
-serde_json = "1.0"
-thiserror = { workspace = true }
-serde-json-wasm = "1.0.1"
 hex = "0.4.3"
-sha2 = "0.10.8"
 
 [dev-dependencies]
 anyhow = "1"
