+++ contracts/contracts/ethereum-analyzer/src/contract.rs
@@ -1,9 +1,10 @@
 use crate::error::ContractError;
 use crate::ethereum;
-use crate::msg::{AnalyzeResult, EthereumAnalyzerResult, ExecuteMsg, QueryMsg};
+use crate::msg::{EthereumAnalyzerResult, ExecuteMsg, QueryMsg};
 use alloy_consensus::SignableTransaction;
 use alloy_primitives::TxKind;
 use bindings::WardenProtocolQuery;
+use contracts_core::msg::AnalyzeResult;
 #[cfg(not(feature = "library"))]
 use cosmwasm_std::entry_point;
 use cosmwasm_std::{Binary, Deps, DepsMut, Empty, Env, MessageInfo, Response, StdError, StdResult};
