+++ contracts/contracts/ethereum-analyzer/src/msg.rs
@@ -7,28 +7,6 @@ pub enum ExecuteMsg {
 }
 
 #[cw_serde]
-pub struct AnalyzeResult<T> {
-    pub data_for_signing: Option<Binary>,
-    pub result: T,
-}
-
-impl<T> AnalyzeResult<T> {
-    pub fn new(result: T) -> AnalyzeResult<T> {
-        AnalyzeResult {
-            data_for_signing: None,
-            result,
-        }
-    }
-
-    pub fn new_with_data(data: Binary, result: T) -> AnalyzeResult<T> {
-        AnalyzeResult {
-            data_for_signing: Some(data),
-            result,
-        }
-    }
-}
-
-#[cw_serde]
 pub struct EthereumAnalyzerResult {
     pub to: String,
     pub value: String,
